Add authorization to product management/profile view

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import get_object_or_404, render, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -87,10 +88,18 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     Add a new product to the store
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request,
+            "Only superusers may add new products."
+            )
+        redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -116,10 +125,18 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit an existing product
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request,
+            "Only superusers may edit products."
+            )
+        redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -147,10 +164,18 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete an existing product
     """
+    if not request.user.is_superuser:
+        messages.error(
+            request,
+            "Only superusers may delete products."
+            )
+        redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
